+++ app/helpers/getObjectData.ts
@@ -1,0 +1,6 @@
+export function getObjectData(objectDefinition) {
+  return {
+    category: objectDefinition.category,
+    type: objectDefinition.type,
+  };
+}
